refactor : 서비스 로직 인터페이스 삭제
